Change pw fn works & closes component on submit
