+++ Scout/src/App.jsx
@@ -7,6 +7,7 @@ import Footer from "./components/Footer";
 import JobPostingsPage from "./components/JobPostingsPage";
 import RegisterUser from "./components/RegisterUser";
 import KanbanBoard from "./components/KanbanBoard";
+import ChangePassword from "./components/ChangePassword";
 
 
 function App() {
@@ -24,6 +25,42 @@ function App() {
 	const [theme, setTheme] = useState("dark");
 	const [registerPage, setRegisterPage] = useState(false);
 	const [userName, setUserName] = useState("");
+	const [changePassword, setChangePassword] = useState(false)
+	const [currentPassword, setCurrentPassword] = useState("");
+	const [newPassword, setNewPassword] = useState("");
+
+const handleChangePassword = () => {
+		setChangePassword(true)
+	}
+	const handlePasswordChange = async () => {
+		try {
+			let response = await fetch(
+				`https://scouttestserver.onrender.com/api/changePassword`,
+				{
+					method: "PATCH",
+					headers: {
+						"Content-Type": "application/json",
+					},
+					body: JSON.stringify({
+						user_id: userId,
+						oldPassword: currentPassword,
+						newPassword: newPassword,
+					}),
+				}
+			);
+
+			if (response.ok) {
+				let resData = await response.json();
+				console.log("password was changed", resData);
+			} else {
+				console.log("failed to change password");
+			}
+		} catch (error) {
+			console.log(error.stack);
+		}
+	};
+
+	
 
 	useEffect(() => {
 		const getUser = async () => {
@@ -105,8 +142,12 @@ function App() {
 						userType={userType}
 						changeRegisterPage={changeRegisterPage}
 						userName={userName}
+						setChangePassword={setChangePassword}
+						handleChangePassword={handleChangePassword}
 					/>
-					{registerPage ? (
+					
+					{changePassword ? <ChangePassword userId={userId} handlePasswordChange={handlePasswordChange} currentPassword={currentPassword} newPassword={newPassword} setCurrentPassword={setCurrentPassword} setNewPassword={setNewPassword} setChangePassword={setChangePassword}/> : 
+					registerPage ? (
 						<RegisterUser />
 					) : jobPosting ? (
 						<JobPostingsPage
